shifted the devise model to User not individual
The theory here is that anyone should be able to sign up or unsignup
from the site and get minimal permissions. If there is a record in the
database with the same email or if the account is manually linked to a
student, then it gets more permissions. This way, user accounts link to
registrar accounts instead of being one and the same.